Render UsertTypes with help texts
